Adds IP blocklist model
