+++ bookwyrm/middleware/__init__.py
@@ -1,3 +1,3 @@
 """ look at all this nice middleware! """
-from timezone_middleware import TimezoneMiddleware
-from ip_middleware import IPBlocklistMiddleware
+from .timezone_middleware import TimezoneMiddleware
+from .ip_middleware import IPBlocklistMiddleware
