+++ bookwyrm/models/__init__.py
@@ -26,8 +26,8 @@ from .import_job import ImportJob, ImportItem
 
 from .site import SiteSettings, SiteInvite
 from .site import PasswordReset, InviteRequest
-from .site import EmailBlocklist
 from .announcement import Announcement
+from .antispam import EmailBlocklist, IPBlocklist
 
 cls_members = inspect.getmembers(sys.modules[__name__], inspect.isclass)
 activity_models = {
